feat: merge software and hardware info pages into single information page
