+++ lact-gui/i18n/it/lact_gui.ftl
@@ -6,9 +6,8 @@ cache-info = Informazioni Cache
 watt = W
 power-usage = Consumo Energetico
 power-profile-mode = Modalità Profilo Di Alimentazione:
-software-page = Software
-hardware-info = Informazioni Hardware
 system-section = Sistema
+hardware-section = Hardware
 lact-gui = GUI LACT
 lact-daemon = Demone LACT
 instance = Istanza
@@ -100,7 +99,7 @@ nvidia-oc-description =
 missing-stat = N/D
 performance-level-auto = Automatico
 performance-level-manual = Manuale
-info-page = Informazioni
+information-page = Informazioni
 amd-oc = AMD Overclocking
 enable-vram-locked-clocks = Abilita Frequenze Bloccate VRAM
 enable-gpu-locked-clocks = Abilita Frequenze Bloccate GPU
