+++ lact-gui/src/app/pages/info_page/vulkan/mod.rs
