+++ lact-gui/src/app/pages/mod.rs
@@ -1,7 +1,6 @@
 pub mod crash_page;
 pub mod info_page;
 pub mod oc_page;
-pub mod software_page;
 pub mod thermals_page;
 
 use lact_schema::{DeviceInfo, DeviceStats};
